pyproject.toml
[tool.poetry]
name = "catalog-project"
version = "0.1.0"
description = ""
authors = ["Toreakhmet <s.vasilovich@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
